@@ -19,9 +19,9 @@ version: 1.0.0+1
 
 environment:
   #dart sdk
-  sdk: 2.8.0-dev.12.0.flutter-9983424a3c
+  sdk: 2.14.4
   #flutter sdk
-  flutter: 1.15.18
+  flutter: 2.5.3
 
 dependencies:
   flutter:
@@ -30,31 +30,31 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   #state管理
-  provider: ^4.0.1
+  provider: ^6.0.1
   #网络请求库
-  dio: ^3.0.3
+  dio: ^4.0.4
   #eventbus事件总线
-  event_bus:  ^1.1.0
+  event_bus:  ^2.0.0
   #文本路劲库
-  path_provider: ^1.3.0
+  path_provider: ^2.0.7
   #权限库
-  permission_handler: 3.2.2
+#  permission_handler: ^8.3.0
   #webView
-  webview_flutter: ^0.3.9+1
+  webview_flutter: ^2.3.1
   #toast
-  fluttertoast: ^3.1.3
-  android_intent: 0.3.0
+  fluttertoast: ^8.0.8
+  android_intent: ^2.0.2
   #sp
-  shared_preferences: 0.5.3+4
-  connectivity: 0.4.4
+  shared_preferences: ^2.0.9
+  connectivity: ^3.0.6
   # json解析库
-  json_annotation: ^3.0.0
+  json_annotation: ^4.4.0
   # log输出库
-  logger: ^0.8.2
+  logger: ^1.1.0
 
 dev_dependencies:
-  json_serializable: ^3.2.3
-  build_runner: ^1.7.1
+  json_serializable: ^6.1.0
+  build_runner: ^2.1.5
   flutter_test:
     sdk: flutter
 
