@@ -12,20 +12,9 @@ const NOT_TIP_KEY = "noTip";
  * on 2019/3/23.
  */
 class ErrorInterceptors extends InterceptorsWrapper {
-  final Dio _dio;
+  final Dio? _dio;
 
   ErrorInterceptors(this._dio);
 
-  @override
-  onRequest(RequestOptions options) async {
-    //没有网络
-    var connectivityResult = await (new Connectivity().checkConnectivity());
-    if (connectivityResult == ConnectivityResult.none) {
-      return _dio.resolve(new ResultData(
-          Code.errorHandleFunction(Code.NETWORK_ERROR, "", false),
-          false,
-          Code.NETWORK_ERROR));
-    }
-    return options;
-  }
+
 }
