@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 class SizeRoute extends PageRouteBuilder {
   final Widget widget;
 
-  SizeRoute({this.widget})
+  SizeRoute({required this.widget})
       : super(
           pageBuilder: (
             BuildContext context,
@@ -30,7 +30,7 @@ class SizeRoute extends PageRouteBuilder {
 class NoAnimationRoute extends PageRouteBuilder {
   final Widget widget;
 
-  NoAnimationRoute({this.widget})
+  NoAnimationRoute({required this.widget})
       : super(
           pageBuilder: (
             BuildContext context,
